moved game to interface folder
