+++ source/battle-board/battle-board/models/interfaces/game.ts
